Adding helper class SnapshotCoordinator
The SnapshotCoordinator is used to overcome some dangerous behavior
when using Saved Game API (aka Snapshots).  The problem is caused by the
not having way to prevent having the same file opened multiple overlapping times.
Also there is no way to stop the incorrect reusing of the Snapshot contained
in the OpenSnapshotResult object.  It should be used once to write or be closed.
Multiple commits using the same snapshot will result in unrecoverable conflicts.

This class is used to encapsulate the access to the Snapshots API and
enforce these semantics of exclusive file access and one-time committing of
metadata.

Change-Id: I18ac550979c038ba4c17c88d48ad304505f95ba5